RBC/MeshIO: move writing functions and various refactor/tidy
- move writing functions into MeshIO as members of MeshIO
- move MeshIO into its own files
- move Cell.impl.c unpleasantness into CellEnergy
- same for Facet.impl.cc -> Facet.cc
- a bit of include what you use work for mesh related TUs
- adapt SimMaster and unit tests to cope
