+++ Code/redblood/CMakeLists.txt
@@ -10,7 +10,9 @@ if (HEMELB_BUILD_RBC)
 
   add_library(hemelb_redblood
     RBCConfig.cc
-    Mesh.cc CellBase.cc Cell.cc Interpolation.cc
+    Mesh.cc MeshIO.cc
+    CellBase.cc Cell.cc CellEnergy.cc Facet.cc
+    Interpolation.cc
     CellCell.cc FlowExtension.cc FaderCell.cc xmlIO.cc RBCInserter.cc
     VertexBag.cc Borders.cc
     WallCellPairIterator.cc
