+++ Code/redblood/Cell.h
@@ -11,15 +11,15 @@
 #include <set>
 #include <utility>
 
-#include "redblood/Mesh.h"
 #include "redblood/CellBase.h"
-#include "redblood/Node2Node.h"
 #include "units.h"
 
 namespace hemelb
 {
   namespace redblood
   {
+    class Mesh;
+
     //! Deformable cell for which energy and forces can be computed
     class Cell : public CellBase
     {
@@ -145,17 +145,6 @@ namespace hemelb
         and (not std::is_pod<Cell::Moduli>::value),
         "Explicit type characteristics"
     );
-
-    //! Write cell-mesh to file in Timm's format
-    void writeMesh(std::ostream &, std::shared_ptr<CellBase const>, util::UnitConverter const&);
-    //! Write cell-mesh to file in Timm's format
-    void writeMesh(std::string const&, std::shared_ptr<CellBase const>, util::UnitConverter const&);
-    //! Write cell-mesh to file in VTK XML format including individual forces
-    void writeVTKMeshWithForces(std::ostream &, std::shared_ptr<Cell const>,
-                                util::UnitConverter const&);
-    //! Write cell-mesh to file in VTK XML format including individual forces
-    void writeVTKMeshWithForces(std::string const &, std::shared_ptr<Cell const>,
-                                util::UnitConverter const&);
 
   }
 } // namespace hemelb::redblood
