+++ Code/redblood/Mesh.cc
@@ -12,10 +12,8 @@
 #include <vtkPolyData.h>
 #include <vtkPolyDataNormals.h>
 #include <vtkSmartPointer.h>
-#include <vtkXMLPolyDataReader.h>
 
 #include "redblood/Mesh.h"
-#include "redblood/VTKError.h"
 #include "util/fileutils.h"
 #include "log/Logger.h"
 #include "Exception.h"
@@ -25,306 +23,10 @@ namespace hemelb
 {
   namespace redblood
   {
-
-    static std::shared_ptr<MeshData> read_krueger_mesh(std::istream &stream, bool fixFacetOrientation)
-    {
-      log::Logger::Log<log::Debug, log::Singleton>("Reading red blood cell from stream");
-
-      std::string line;
-
-      // Drop header
-      for (int i(0); i < 4; ++i)
-      {
-        std::getline(stream, line);
-      }
-
-      // Number of vertices
-      unsigned int num_vertices;
-      stream >> num_vertices;
-
-      // Create Mesh data
-      std::shared_ptr<MeshData> result(new MeshData);
-      result->vertices.resize(num_vertices);
-
-      // Then read in first and subsequent lines
-      MeshData::Facet::value_type offset;
-      stream >> offset >> result->vertices[0].x >> result->vertices[0].y >> result->vertices[0].z;
-      log::Logger::Log<log::Trace, log::Singleton>("Vertex 0 at %d, %d, %d",
-                                                   result->vertices[0].x,
-                                                   result->vertices[0].y,
-                                                   result->vertices[0].z);
-
-      for (unsigned int i(1), index(0); i < num_vertices; ++i)
-      {
-        stream >> index >> result->vertices[i].x >> result->vertices[i].y >> result->vertices[i].z;
-        // No gaps in vertex index list
-        assert(index == (i + offset));
-        log::Logger::Log<log::Trace, log::Singleton>("Vertex %i at %d, %d, %d",
-                                                     i,
-                                                     result->vertices[i].x,
-                                                     result->vertices[i].y,
-                                                     result->vertices[i].z);
-      }
-
-      // Drop mid-file headers
-      for (int i(0); i < 3; ++i)
-      {
-        std::getline(stream, line);
-      }
-
-      // Read facet indices
-      unsigned int num_facets;
-      MeshData::Facet indices;
-      stream >> num_facets;
-      result->facets.resize(num_facets);
-
-      for (unsigned int i(0), dummy(0); i < num_facets; ++i)
-      {
-        stream >> dummy >> dummy >> dummy >> dummy >> dummy >> dummy >> indices[0] >> indices[1]
-            >> indices[2];
-        result->facets[i][0] = indices[0] - offset;
-        result->facets[i][1] = indices[1] - offset;
-        result->facets[i][2] = indices[2] - offset;
-        log::Logger::Log<log::Trace, log::Singleton>("Facet %i with %i, %i, %i",
-                                                     i,
-                                                     indices[0] - offset,
-                                                     indices[1] - offset,
-                                                     indices[2] - offset);
-      }
-
-      log::Logger::Log<log::Debug, log::Singleton>("Read %i vertices and %i triangular facets",
-                                                   num_vertices,
-                                                   num_facets);
-
-      if (fixFacetOrientation)
-      {
-        orientFacets(*result);
-      }
-      return result;
-    }
-
-    auto KruegerMeshIO::readFile(std::string const &filename, bool fixFacetOrientation) const -> MeshPtr {
-      if (!util::file_exists(filename.c_str()))
-        throw Exception() << "Red-blood-cell mesh file '" << filename.c_str() << "' does not exist";
-
-      log::Logger::Log<log::Debug, log::Singleton>("Reading red blood cell from %s",
-                                                   filename.c_str());
-      // Open file if it exists
-      auto stream = std::ifstream{filename};
-      return read_krueger_mesh(stream, fixFacetOrientation);
-    }
-
-    auto KruegerMeshIO::readString(std::string const &data, bool fixFacetOrientation) const -> MeshPtr {
-      auto stream = std::istringstream{data};
-      return read_krueger_mesh(stream, fixFacetOrientation);
-    }
-
-    auto VTKMeshIO::readUnoriented(Mode mode, std::string const &file_or_data) const -> std::tuple<MeshPtr, PolyDataPtr>
-    {
-      // Read in VTK polydata object
-      auto reader = ErrThrower<vtkXMLPolyDataReader>::New();
-      switch (mode) {
-      case Mode::file:
-	log::Logger::Log<log::Debug, log::Singleton>("Reading red blood cell from VTK polydata file");
-	reader->ReadFromInputStringOff();
-	reader->SetFileName(file_or_data.c_str());
-	break;
-      case Mode::string:
-	log::Logger::Log<log::Debug, log::Singleton>("Reading red blood cell from VTK polydata string");
-	reader->ReadFromInputStringOn();
-	reader->SetInputString(file_or_data);
-	break;
-      }
-
-      reader->Update();
-
-      vtkSmartPointer<vtkPolyData> polydata(reader->GetOutput());
-
-      // Number of vertices
-      unsigned int num_vertices = polydata->GetNumberOfPoints();
-
-      // Create Mesh data
-      std::shared_ptr<MeshData> result(new MeshData);
-      result->vertices.resize(num_vertices);
-
-      // Then read in first and subsequent lines
-      for (unsigned int i(0); i < num_vertices; ++i)
-      {
-        double* point_coord = polydata->GetPoints()->GetPoint(i);
-        result->vertices[i].x = point_coord[0];
-        result->vertices[i].y = point_coord[1];
-        result->vertices[i].z = point_coord[2];
-
-        log::Logger::Log<log::Trace, log::Singleton>("Vertex %i at %e, %e, %e",
-                                                     i,
-                                                     result->vertices[i].x,
-                                                     result->vertices[i].y,
-                                                     result->vertices[i].z);
-      }
-
-      // Read facet indices
-      unsigned int num_facets = polydata->GetNumberOfCells();;
-      result->facets.resize(num_facets);
-
-      for (unsigned int i(0); i < num_facets; ++i)
-      {
-        vtkCell* triangle = polydata->GetCell(i);
-        assert(triangle->GetCellType() == VTK_TRIANGLE);
-        result->facets[i][0] = triangle->GetPointId(0);
-        result->facets[i][1] = triangle->GetPointId(1);
-        result->facets[i][2] = triangle->GetPointId(2);
-        log::Logger::Log<log::Trace, log::Singleton>("Facet %i with %i, %i, %i",
-                                                     i,
-                                                     result->facets[i][0],
-                                                     result->facets[i][1],
-                                                     result->facets[i][2]);
-      }
-
-      log::Logger::Log<log::Debug, log::Singleton>("Read %i vertices and %i triangular facets",
-                                                   num_vertices,
-                                                   num_facets);
-
-      return std::make_tuple(result, polydata);
-    }
-
-    static MeshIO::MeshPtr VTK_mesh_read_impl(VTKMeshIO const & self, VTKMeshIO::Mode m, std::string const &file_or_data, bool fixFacetOrientation) {
-	MeshIO::MeshPtr meshData;
-	VTKMeshIO::PolyDataPtr polyData;
-	std::tie(meshData, polyData) = self.readUnoriented(m, file_or_data);
-
-	if (fixFacetOrientation) {
-	  unsigned numSwapped = orientFacets(*meshData, *polyData);
-	  log::Logger::Log<log::Debug, log::Singleton>("Swapped %d facets", numSwapped);
-	}
-
-	return meshData;
-      }
-
-      auto VTKMeshIO::readFile(std::string const &filename, bool fixFacetOrientation) const -> MeshPtr {
-	return VTK_mesh_read_impl(*this, Mode::file, filename, fixFacetOrientation);
-      }
-
-      auto VTKMeshIO::readString(std::string const &data, bool fixFacetOrientation) const -> MeshPtr {
-	return VTK_mesh_read_impl(*this, Mode::string, data, fixFacetOrientation);
-      }
-
-    void writeMesh(std::string const &filename, MeshData const &data, util::UnitConverter const& c)
-    {
-      log::Logger::Log<log::Debug, log::Singleton>("Writing red blood cell to %s",
-                                                   filename.c_str());
-      std::ofstream file(filename.c_str());
-      assert(file.is_open());
-      writeMesh(file, data, c);
-    }
-
-    void writeVTKMesh(std::string const &filename, MeshData const &data,
-                      util::UnitConverter const& converter)
-    {
-      log::Logger::Log<log::Debug, log::Singleton>("Writing red blood cell to %s",
-                                                   filename.c_str());
-      std::ofstream file(filename.c_str());
-      assert(file.is_open());
-      writeVTKMesh(file, data, converter);
-    }
-
-    void writeMesh(std::ostream &stream, MeshData const &data, util::UnitConverter const& converter)
-    {
-      writeMesh(stream, data.vertices, data.facets, converter);
-    }
-
-    void writeMesh(std::ostream &stream, MeshData::Vertices const &vertices,
-                   MeshData::Facets const & facets, util::UnitConverter const& converter)
-    {
-      // Write Header
-      stream << "$MeshFormat\n2 0 8\n$EndMeshFormat\n" << "$Nodes\n" << vertices.size()
-          << "\n";
-
-      typedef MeshData::Vertices::const_iterator VertexIterator;
-      VertexIterator i_vertex = vertices.begin();
-      VertexIterator const i_vertex_end = vertices.end();
-
-      for (unsigned i(1); i_vertex != i_vertex_end; ++i_vertex, ++i)
-      {
-        auto const vertex = converter.ConvertPositionToPhysicalUnits(*i_vertex);
-        stream << i << " " << vertex[0] << " " << vertex[1] << " " << vertex[2] << "\n";
-      }
-
-      stream << "$EndNode\n" << "$Elements\n" << facets.size() << "\n";
-
-      typedef MeshData::Facets::const_iterator FacetIterator;
-      FacetIterator i_facet = facets.begin();
-      FacetIterator const i_facet_end = facets.end();
-
-      for (unsigned i(1); i_facet != i_facet_end; ++i_facet, ++i)
-        stream << i << " 1 2 3 4 5 " << (*i_facet)[0] + 1 << " " << (*i_facet)[1] + 1 << " "
-            << (*i_facet)[2] + 1 << "\n";
-
-      stream << "$EndElement\n";
-    }
-
-    void writeVTKMesh(std::ostream &stream, MeshData const &data, util::UnitConverter const &conv)
-    {
-      writeVTKMesh(stream, data.vertices, data.facets, conv);
-    }
-    void writeVTKMesh(std::ostream &stream, MeshData::Vertices const &vertices,
-                      MeshData::Facets const &facets, util::UnitConverter const & converter,
-                      PointScalarData pointScalarData)
-    {
-      // Write Header
-      stream << "<?xml version=\"1.0\"?>\n"
-          << "<VTKFile type=\"PolyData\" version=\"0.1\" byte_order=\"LittleEndian\">\n"
-          << "  <PolyData>\n" << "    <Piece NumberOfPoints=\"" << vertices.size()
-          << "\" NumberOfPolys=\"" << facets.size() << "\">\n" << "      <Points>\n"
-          << "        <DataArray NumberOfComponents=\"3\" type=\"Float32\">\n";
-
-      typedef MeshData::Vertices::const_iterator VertexIterator;
-      VertexIterator i_vertex = vertices.begin();
-      VertexIterator const i_vertex_end = vertices.end();
-
-      for (unsigned i(1); i_vertex != i_vertex_end; ++i_vertex, ++i)
-      {
-        auto const vertex = converter.ConvertPositionToPhysicalUnits(*i_vertex);
-        stream << vertex[0] << " " << vertex[1] << " " << vertex[2] << " ";
-      }
-
-      stream << "\n        </DataArray>\n" << "      </Points>\n";
-
-      if (pointScalarData.size() > 0)
-      {
-        stream << "      <PointData Scalar=\"scalar_fields\">\n";
-
-        for (auto field : pointScalarData)
-        {
-          stream << "        <DataArray type=\"Float32\" Name=\"" << field.first << "\">\n";
-          stream << "        ";
-          typedef PointScalarData::value_type::second_type::value_type ScalarFieldType;
-          std::ostream_iterator<ScalarFieldType> stream_iterator(stream, " ");
-          std::copy(field.second.begin(), field.second.end(), stream_iterator);
-          stream << "\n        </DataArray>\n";
-        }
-        stream << "      </PointData>\n";
-      }
-
-      stream << "      <Polys>\n" << "        <DataArray type=\"Int32\" Name=\"connectivity\">\n";
-
-      typedef MeshData::Facets::const_iterator FacetIterator;
-      FacetIterator i_facet = facets.begin();
-      FacetIterator const i_facet_end = facets.end();
-
-      for (; i_facet != i_facet_end; ++i_facet)
-        stream << (*i_facet)[0] << " " << (*i_facet)[1] << " " << (*i_facet)[2] << " ";
-
-      stream << "\n        </DataArray>\n"
-          << "        <DataArray type=\"Int32\" Name=\"offsets\">\n";
-
-      for (unsigned i(0); i < facets.size(); ++i)
-      {
-        stream << (i + 1) * 3 << " ";
-      }
-
-      stream << "\n        </DataArray>\n" << "      </Polys>\n";
-      stream << "    </Piece>\n  </PolyData>\n</VTKFile>\n";
-    }
+    // Check that the condition in Mesh.h is true.
+    // TODO: fix as described in header
+    static_assert(std::is_same<IdType, vtkIdType>::value,
+		  "hemelb::redblood::IdType must be the same as vtkIdType");
 
     LatticePosition barycenter(MeshData::Vertices const &vertices)
     {
