+++ Code/redblood/xmlIO.cc
@@ -7,12 +7,14 @@
 #include "configuration/SimConfig.h"
 #include "io/xml/XmlAbstractionLayer.h"
 #include "redblood/stencil.h"
+#include "redblood/xmlIO.h"
 #include "redblood/Cell.h"
 #include "redblood/FaderCell.h"
-#include "redblood/RBCInserter.h"
+#include "redblood/MeshIO.h"
+#include "redblood/Node2Node.h"
 #include "redblood/RBCConfig.h"
+#include "redblood/RBCInserter.h"
 #include "redblood/FlowExtension.h"
-#include "redblood/xmlIO.h"
 #include "Traits.h"
 
 namespace hemelb
