+++ Code/unittests/redblood/CellVelocityInterpolTests.h
@@ -8,7 +8,6 @@
 
 #include <cppunit/TestFixture.h>
 #include "redblood/GridAndCell.h"
-#include "redblood/Facet.impl.cc"
 #include "unittests/redblood/Fixtures.h"
 
 namespace hemelb
