+++ Code/unittests/redblood/EnergyTests.h
@@ -8,6 +8,7 @@
 
 #include <cppunit/TestFixture.h>
 #include "resources/Resource.h"
+#include "redblood/CellEnergy.h"
 #include "unittests/redblood/Fixtures.h"
 
 namespace hemelb
